requirements.txt
aggdraw==1.3.18.post0
altgraph==0.17.4
attrs==24.2.0
docopt==0.6.2
greenlet==3.1.1
imagecodecs==2024.9.22
imageio==2.35.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
macholib==1.16.3
modulegraph==0.19.6
networkx==3.3
numpy==1.26.4
opencv-python==4.6.0.66
packaging==24.1
pillow==11.0.0
psd-tools==1.9.34
py2app==0.28.8
pyproject-toml==0.0.10
PyQt5==5.15.10
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
rawpy==0.23.1
referencing==0.35.1
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
SQLAlchemy==2.0.35
tifffile==2023.12.9
toml==0.10.2
typing_extensions==4.12.2
